Align browser engine code with architecture policy
