+++ Packages/macOS/CmuxBrowser/Sources/CmuxBrowser/Engine/Chromium/CDPEventSubscriber.swift
@@ -1,0 +1,4 @@
+struct CDPEventSubscriber {
+    let sessionID: String
+    let continuation: AsyncStream<CDPEvent>.Continuation
+}
