+++ Packages/macOS/CmuxBrowser/Sources/CmuxBrowser/Engine/Chromium/CDPPendingRequest.swift
@@ -1,0 +1,4 @@
+struct CDPPendingRequest {
+    let continuation: CheckedContinuation<CDPJSONValue, any Error>
+    let timeoutTask: Task<Void, Never>
+}
