+++ Packages/macOS/CmuxBrowser/Sources/CmuxBrowser/Engine/Chromium/CDPWebSocketTransport.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+/// Sendable transport boundary for the browser-level DevTools WebSocket.
+protocol CDPWebSocketTransport: Sendable {
+    func resume()
+    func send(_ data: Data) async throws
+    func receive() async throws -> Data
+    func cancel()
+}
