+++ Packages/macOS/CmuxBrowser/Sources/CmuxBrowser/Engine/Chromium/ChromiumIsolatedWorldConfiguration.swift
@@ -1,0 +1,11 @@
+struct ChromiumIsolatedWorldConfiguration {
+    let frameID: String
+
+    var parameters: [String: CDPJSONValue] {
+        [
+            "frameId": .string(frameID),
+            "worldName": .string("cmux.browser.automation"),
+            "grantUniversalAccess": .bool(true),
+        ]
+    }
+}
