+++ Packages/macOS/CmuxBrowser/Sources/CmuxBrowser/Engine/Chromium/ChromiumScreencastTransition.swift
@@ -1,0 +1,9 @@
+struct ChromiumScreencastTransition {
+    let isViewportVisible: Bool
+    let isScreencastActive: Bool
+
+    var method: String? {
+        guard isViewportVisible != isScreencastActive else { return nil }
+        return isViewportVisible ? "Page.startScreencast" : "Page.stopScreencast"
+    }
+}
