+++ Packages/macOS/CmuxBrowser/Sources/CmuxBrowser/Engine/Chromium/ChromiumViewportGestureTransition.swift
@@ -1,0 +1,4 @@
+enum ChromiumViewportGestureTransition: Equatable {
+    case began(String)
+    case ended(String)
+}
