+++ Packages/macOS/CmuxBrowser/Sources/CmuxBrowser/Engine/Chromium/ChromiumViewportInputCoalescingKind.swift
@@ -1,0 +1,4 @@
+enum ChromiumViewportInputCoalescingKind: Equatable {
+    case mouseMove
+    case mouseWheel
+}
