+++ Packages/macOS/CmuxBrowser/Sources/CmuxBrowser/Engine/Chromium/URLSessionCDPWebSocketTransport.swift
@@ -1,0 +1,39 @@
+import Foundation
+
+/// Bridges Foundation's thread-safe WebSocket session into the CDP transport boundary.
+// URLSession and URLSessionWebSocketTask own synchronization; this wrapper keeps only immutable references.
+final class URLSessionCDPWebSocketTransport: CDPWebSocketTransport, @unchecked Sendable {
+    private let session: URLSession
+    private let task: URLSessionWebSocketTask
+
+    init(url: URL) {
+        let configuration = URLSessionConfiguration.ephemeral
+        let session = URLSession(configuration: configuration)
+        self.session = session
+        self.task = session.webSocketTask(with: url)
+    }
+
+    func resume() {
+        task.resume()
+    }
+
+    func send(_ data: Data) async throws {
+        try await task.send(.data(data))
+    }
+
+    func receive() async throws -> Data {
+        switch try await task.receive() {
+        case .data(let data):
+            return data
+        case .string(let value):
+            return Data(value.utf8)
+        @unknown default:
+            return Data()
+        }
+    }
+
+    func cancel() {
+        task.cancel(with: .goingAway, reason: nil)
+        session.invalidateAndCancel()
+    }
+}
