+++ Packages/macOS/CmuxBrowser/Sources/CmuxBrowser/Engine/WebKitBrowserEngineCookieError.swift
@@ -1,0 +1,4 @@
+/// A WebKit cookie could not be represented as an `HTTPCookie`.
+enum WebKitBrowserEngineCookieError: Error {
+    case invalidPayload
+}
